+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Fix:
+- Fix Legacy Carousel width.
+
 ## [2.89.0] - 2020-08-28
 
 Features: